Add default for Builder limit
If the limit is not explicitly set on the Builder then we get the error, "Devloops\LaravelTypesense\Engines\TypesenseSearchEngine::buildSearchParams(): Argument #3 ($perPage) must be of type int, null given". Here's a fix that works but I'd understand if you have another way that works. I hardcoded the limit to 10 because that's the Typesense default for the per_page argument: https://typesense.org/docs/0.21.0/api/documents.html#arguments.